[ENG-1810] chore: move dynamic field mapping to optional section
