+++ src/components/Configure/content/fields/FieldMappings/RequiredFieldMappings.tsx
@@ -1,9 +1,7 @@
 import { useMemo } from 'react';
-import { IntegrationFieldMapping } from '@generated/api/src';
 
 import { ErrorBoundary, useErrorState } from 'context/ErrorContextProvider';
 import { FormControl } from 'src/components/form/FormControl';
-import { useInstallIntegrationProps } from 'src/context/InstallIntegrationContextProvider';
 
 import { useSelectedConfigureState } from '../../useSelectedConfigureState';
 import { FieldHeader } from '../FieldHeader';
@@ -13,7 +11,6 @@ import { setFieldMapping } from './setFieldMapping';
 
 export function RequiredFieldMappings() {
   const { selectedObjectName, configureState, setConfigureState } = useSelectedConfigureState();
-  const { fieldMapping } = useInstallIntegrationProps();
   const { isError, removeError } = useErrorState();
 
   const onSelectChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
@@ -35,28 +32,10 @@ export function RequiredFieldMappings() {
     }
   };
 
-  const integrationFieldMappings = useMemo(() => {
-    // 1. Extract required map fields from configureState
-    const requiredFieldMappings = configureState?.read?.requiredMapFields || [];
-
-    // 2. Extract dynamic field mappings for the selected object name from the fieldMapping object if it exists
-    const dynamicFieldMappings = selectedObjectName && fieldMapping
-      ? Object.values(fieldMapping[selectedObjectName] || {})
-        .flat()
-        .filter((mapping) => !mapping.fieldName)
-      : [];
-
-    // 3. Combine dynamic field mappings with the required map fields from configureState
-    const combinedFieldMappings = requiredFieldMappings.concat(dynamicFieldMappings)
-      // 4. Remove duplicates based on mapToName and keep the latest item
-      .reduce((acc, item) => {
-        const existingItem = acc.find((i) => i.mapToName === item.mapToName);
-        if (existingItem) return acc.map((i) => (i.mapToName === item.mapToName ? item : i));
-        return acc.concat(item);
-      }, new Array<IntegrationFieldMapping>());
-
-    return combinedFieldMappings;
-  }, [configureState, fieldMapping, selectedObjectName]);
+  const integrationFieldMappings = useMemo(
+    () => configureState?.read?.requiredMapFields || [],
+    [configureState],
+  );
 
   return integrationFieldMappings?.length ? (
     <>
